Fix bug where some arguments would not be recognized as invalid
Since data_arg_index only checked if the data argument started with the analyzed argument, arguments that were only prefixes of data arguments (e.g. '-ma', prefix of the valid argument '-mac') would be recognized as data arguments. The parsing would still stop because the next argument would not be considered valid as an option of the misinterpreted previous argument, but the error output would not be expressive of what was wrong in the arguments
Now data_arg_index checks if both the data argument and the argument analyzed start with each other, ensuring that the 'prefix only' arguments that caused the aforementioned problems are recognized as invalid